Fix implicit casts of numbers & int multiplication (#1002)
* Cast or replace with appropriate type to prevent overflow or data loss
* Use long literal for int multiplication where possible

+++ core/src/main/java/org/apache/accumulo/core/clientImpl/TabletServerBatchWriter.java
@@ -1002,7 +1002,7 @@ public class TabletServerBatchWriter {
   private static class MutationSet {
 
     private final HashMap<TableId,List<Mutation>> mutations;
-    private int memoryUsed = 0;
+    private long memoryUsed = 0;
 
     MutationSet() {
       mutations = new HashMap<>();
@@ -1050,7 +1050,7 @@ public class TabletServerBatchWriter {
       }
     }
 
-    public int getMemoryUsed() {
+    public long getMemoryUsed() {
       return memoryUsed;
     }
